requirements.txt
Django>=5,<6

# Episode 2: Django Forms
django-widget-tweaks>=1.5,<2
django-crispy-forms>=2.1,<3
crispy-bulma>=0.11<1
django-htmx>=1.17.3,<2

# Episode 5: DRF & Django Ninja
djangorestframework==3.15.2
django-ninja==1.3.0
